+++ src/main/java/com/google/api/generator/engine/writer/ImportWriterVisitor.java
@@ -17,6 +17,7 @@ package com.google.api.generator.engine.writer;
 import com.google.api.generator.engine.ast.AnnotationNode;
 import com.google.api.generator.engine.ast.AnonymousClassExpr;
 import com.google.api.generator.engine.ast.ArithmeticOperationExpr;
+import com.google.api.generator.engine.ast.ArrayExpr;
 import com.google.api.generator.engine.ast.AssignmentExpr;
 import com.google.api.generator.engine.ast.AssignmentOperationExpr;
 import com.google.api.generator.engine.ast.AstNodeVisitor;
@@ -167,6 +168,11 @@ public class ImportWriterVisitor implements AstNodeVisitor {
     annotation.type().accept(this);
   }
 
+  @Override
+  public void visit(ArrayExpr expr) {
+    expr.type().accept(this);
+  }
+
   /** =============================== EXPRESSIONS =============================== */
   @Override
   public void visit(ValueExpr valueExpr) {
@@ -183,6 +189,7 @@ public class ImportWriterVisitor implements AstNodeVisitor {
   @Override
   public void visit(VariableExpr variableExpr) {
     variableExpr.variable().type().accept(this);
+    annotations(variableExpr.annotations());
     if (variableExpr.exprReferenceExpr() != null) {
       variableExpr.exprReferenceExpr().accept(this);
     }